requirements.txt
# Core ML stack (pinned)
torch==2.9.1+cu128
transformers==4.57.3
peft==0.18.1
bitsandbytes==0.48.2
accelerate==1.12.0

# Data (pinned)
datasets==4.4.1

# System / GPU utils (pinned)
pynvml==13.0.1

# Configuration
pyyaml==6.0.2

# Testing
pytest==8.3.5
pytest-cov==6.0.0

# Logging (optional, for enhanced logging)
# loguru>=0.7.0
